one last big refactor
