+++ lua/codecompanion/acp/methods.lua
@@ -1,0 +1,12 @@
+return {
+  INITIALIZE = "initialize",
+  AUTHENTICATE = "authenticate",
+  SESSION_CANCEL = "session/cancel",
+  SESSION_LOAD = "session/load",
+  SESSION_NEW = "session/new",
+  SESSION_PROMPT = "session/prompt",
+  SESSION_REQUEST_PERMISSION = "session/request_permission",
+  SESSION_UPDATE = "session/update",
+  FS_READ_TEXT_FILE = "fs/read_text_file",
+  FS_WRITE_TEXT_FILE = "fs/write_text_file",
+}
