+++ tests/acp/test_acp.lua
@@ -507,150 +507,4 @@ T["ACP Responses"]["ignores notifications for other sessions"] = function()
   h.eq(result.last, "seen")
 end
 
-T["ACP Responses"]["PromptBuilder cancel sends notification (no id)"] = function()
-  local result = child.lua([[
-    local connection = ACP.new({ adapter = test_adapter })
-    connection.session_id = "test-session-123"
-    connection.methods = { encode = vim.json.encode }
-    connection._state = { next_id = 1 }
-
-    local sent = {}
-    connection._write_to_process = function(self, data)
-      table.insert(sent, vim.trim(data))
-      return true
-    end
-
-    local prompt = connection:prompt({ { role = "user", content = "hi" } })
-    prompt:cancel()
-    local obj = vim.json.decode(sent[#sent])
-    return { has_id = obj.id ~= nil, method = obj.method, sessionId = obj.params.sessionId }
-  ]])
-
-  h.eq(result.has_id, false)
-  h.eq(result.method, "session/cancel")
-  h.eq(result.sessionId, "test-session-123")
-end
-
-T["ACP Responses"]["PromptBuilder"] = function()
-  local result = child.lua([[
-    local connection = ACP.new({ adapter = test_adapter })
-    connection.session_id = "test-session-123"
-    connection.process = { next_id = 10 }
-    connection.methods = { encode = vim.json.encode }
-
-    -- Track handler calls and sent data
-    local handler_calls = {}
-    local sent_data = ""
-
-    connection._write_to_process = function(self, data)
-      sent_data = data
-      return true
-    end
-
-    -- Create prompt with fluent API
-    local prompt = connection:prompt({
-      { role = "user", content = "Test message" }
-    })
-    :on_message_chunk(function(content)
-      table.insert(handler_calls, { type = "message", content = content })
-    end)
-    :on_thought_chunk(function(content)
-      table.insert(handler_calls, { type = "thought", content = content })
-    end)
-    :on_complete(function(reason)
-      table.insert(handler_calls, { type = "complete", reason = reason })
-    end)
-    :with_options({ silent = true })
-
-    -- Send and simulate responses
-    local job = prompt:send()
-    prompt:_handle_session_update({
-      sessionUpdate = "agent_thought_chunk",
-      content = { type = "text", text = "Thinking..." }
-    })
-    prompt:_handle_session_update({
-      sessionUpdate = "agent_message_chunk",
-      content = { type = "text", text = "Hello!" }
-    })
-    prompt:_handle_done()
-
-    return {
-      handler_count = #handler_calls,
-      thought_call = handler_calls[1],
-      message_call = handler_calls[2],
-      complete_call = handler_calls[3],
-      has_shutdown = type(job.shutdown) == "function",
-      sent_request = vim.json.decode(vim.trim(sent_data)),
-    }
-  ]])
-
-  h.eq(result.handler_count, 3)
-  h.eq(result.thought_call.type, "thought")
-  h.eq(result.thought_call.content, "Thinking...")
-  h.eq(result.message_call.type, "message")
-  h.eq(result.message_call.content, "Hello!")
-  h.eq(result.complete_call.type, "complete")
-  h.eq(result.has_shutdown, true)
-  h.eq(result.sent_request.method, "session/prompt")
-  h.eq(result.sent_request.params.sessionId, "test-session-123")
-end
-
-T["ACP Responses"]["PromptBuilder extracts text safely from non-text content"] = function()
-  local result = child.lua([[
-    local connection = ACP.new({ adapter = test_adapter })
-    connection.session_id = "test-session-123"
-
-    local seen = {}
-    local prompt = connection:prompt({ { role = "user", content = "hi" } })
-      :on_message_chunk(function(text) table.insert(seen, text) end)
-      :on_thought_chunk(function(text) table.insert(seen, "T:" .. text) end)
-
-    -- Simulate image and resource_link chunks
-    prompt:_handle_session_update({
-      sessionUpdate = "agent_message_chunk",
-      content = { type = "image", data = "..." }
-    })
-    prompt:_handle_session_update({
-      sessionUpdate = "agent_message_chunk",
-      content = { type = "resource_link", uri = "file:///tmp/x.txt", name = "x" }
-    })
-    prompt:_handle_session_update({
-      sessionUpdate = "agent_thought_chunk",
-      content = { type = "text", text = "thinking" }
-    })
-
-    return seen
-  ]])
-
-  -- Expect placeholders for image/resource_link and proper text for thought
-  h.eq(result[1], "[image]")
-  h.eq(true, result[2]:match("^%[resource: ") ~= nil)
-  h.eq(result[3], "T:thinking")
-end
-
-T["ACP Responses"]["PromptBuilder cancel sends notification (no id)"] = function()
-  local result = child.lua([[
-    local connection = ACP.new({ adapter = test_adapter })
-    connection.session_id = "test-session-123"
-    connection.methods = { encode = vim.json.encode }
-    connection._state = { next_id = 1 } -- ensure state
-
-    local sent = {}
-    connection._write_to_process = function(self, data)
-      table.insert(sent, vim.trim(data))
-      return true
-    end
-
-    local prompt = connection:prompt({ { role = "user", content = "hi" } })
-    prompt:cancel()
-
-    local obj = vim.json.decode(sent[#sent])
-    return { has_id = obj.id ~= nil, method = obj.method, params = obj.params }
-  ]])
-
-  h.eq(result.has_id, false)
-  h.eq(result.method, "session/cancel")
-  h.eq(result.params.sessionId, "test-session-123")
-end
-
 return T
